FormSaver-main/backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
psycopg2-binary
httpx
python-dotenv
python-multipart
